create a components directory and move files into it
