+++ src/components/schema-ui/attribute.js
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
-import { LocationContext } from './contexts/location';
-import { hasSetEntry } from './lib/utils/utils';
-import { SchemaContext } from './contexts/schema';
+import { LocationContext } from '../../contexts/location';
+import { hasSetEntry } from '../../lib/utils/utils';
+import { SchemaContext } from '../../contexts/schema';
 
 const Attribute = ({ attribute, type, includeEnabled }) => {
   const { forPath } = useContext(SchemaContext);
