+++ src/lib/jsonapi-objects/document.js
@@ -1,4 +1,4 @@
-import {Link} from "../../link";
+import {Link} from "../../components/link";
 import ResourceObject from "./resource-object";
 import SchemaParser from "../schema/schema-parser";
