+++ src/lib/jsonapi-objects/resource-object.js
@@ -1,4 +1,4 @@
-import {Link} from "../../link";
+import {Link} from "../../components/link";
 import { extract } from '../utils/utils'
 
 export default class ResourceObject {
